backend/requirements.txt
pyserial==3.5
websockets==15.0.1
